Convert rate to float to be compatible with thumbnails from providers.

+++ dss/mobile/handler.py
@@ -213,7 +213,7 @@ class MediaHandler(socketserver.BaseRequestHandler, object):
             thumb['dir'], self.get_stream_name()
         ) + '.' + thumb['format']
 
-        thumb_rate = str(1. / int(thumb['mobile_interval']))
+        thumb_rate = str(1. / float(thumb['mobile_interval']))
 
         args = ffmpeg.cmd_inputs_outputs(
             '-y -re', [audio_filename, video_filename], '',
